requirements_seg.txt
torch
torchvision
transformers
accelerate
safetensors
opencv-python
numpy
Pillow
